backend/requirements.txt
fastapi 
vicorn
pandas
scikit-learn
joblib
pydantic